Move IsAnalysisScopeMatching
