+++ analyzers/src/SonarAnalyzer.Common/AnalysisContext/SonarAnalysisContext.ScrapYard.cs
@@ -20,7 +20,6 @@
 
 using System.IO;
 using Microsoft.CodeAnalysis.Text;
-using static SonarAnalyzer.Helpers.DiagnosticDescriptorFactory;
 
 namespace SonarAnalyzer;
 
@@ -63,25 +62,6 @@ public partial class SonarAnalysisContext
         else
         {
             return ProjectConfigReader.Empty;
-        }
-    }
-
-    internal static bool IsAnalysisScopeMatching(Compilation compilation, bool isTestProject, bool isScannerRun, IEnumerable<DiagnosticDescriptor> diagnostics)
-    {
-        // We don't know the project type without the compilation so let's run the rule
-        return compilation == null || diagnostics.Any(IsMatching);
-
-        bool IsMatching(DiagnosticDescriptor descriptor)
-        {
-            // MMF-2297: Test Code as 1st Class Citizen is not ready on server side yet.
-            // ScannerRun: Only utility rules and rules with TEST-ONLY scope are executed for test projects for now.
-            // SonarLint & Standalone Nuget: Respect the scope as before.
-            return isTestProject
-                ? ContainsTag(TestSourceScopeTag) && !(isScannerRun && ContainsTag(MainSourceScopeTag) && !ContainsTag(UtilityTag))
-                : ContainsTag(MainSourceScopeTag);
-
-            bool ContainsTag(string tag) =>
-                descriptor.CustomTags.Contains(tag);
         }
     }
